requirements.txt
torch>=2.0
torchvision>=0.15
numpy>=1.24
pillow>=10.0
tifffile>=2023.7
scikit-learn>=1.3
matplotlib>=3.7
seaborn>=0.12
pyyaml>=6.0
tqdm>=4.65
